Actually make pre-mature optimization work
Finaally making it work and preventing the re-rendering of the visual
toast component.

While the context object changes on every notification state change, the
function is memoized and remains the same.

Then I slap the always same function on the ever changing context
object.

Then I pluck the always same function off the always changing context
object. Then the same function is passed to the memoized visual
component. The visual component sees that the function is not changed
and does not re-render.

Of course, this is all premature and the toast comopnents are small and
quick to build (especially since it's visual only). But, I wanted to do
this as a challenge to see if I could do it.